requirements.txt
asgiref==3.5.0
Django==3.2.12
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
gunicorn==20.1.0
psycopg2==2.9.3
PyJWT==2.3.0
python-decouple==3.6
pytz==2021.3
sqlparse==0.4.2
typing_extensions==4.1.1
whitenoise==6.0.0
